refactor: move envelope builder functionality into its own module
