+++ fuzz/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod consts;
+pub mod envelope_builder;
